Refactoring per PR comments
